+++ sunsynk/pysunsynk.py
@@ -4,7 +4,7 @@ import logging
 from typing import Sequence
 from urllib.parse import urlparse
 
-import attr
+import attrs
 from pymodbus import version
 from pymodbus.client import (
     AsyncModbusSerialClient,
@@ -17,77 +17,48 @@ from sunsynk.sunsynk import Sunsynk
 _LOGGER = logging.getLogger(__name__)
 
 
-@attr.define
+@attrs.define
 class pySunsynk(Sunsynk):  # pylint: disable=invalid-name
     """Sunsync Modbus class."""
 
-    port: str = attr.ib(default="/dev/tty0")
-    client: ModbusBaseClient = attr.ib(default=None)
+    client: ModbusBaseClient = None  # type:ignore
+
+    def _new_client(self) -> ModbusBaseClient:
+        """Create a new client."""
+        url = urlparse(f"{self.port}")
+        if url.hostname:
+            host, port = url.hostname, url.port or 502
+            _LOGGER.info("PyModbus %s TCP: %s:%s", version.short(), host, port)
+            return AsyncModbusTcpClient(host=host, port=port)
+
+        _LOGGER.info("PyModbus %s Serial: %s", version.short(), self.port)
+        return AsyncModbusSerialClient(
+            port=self.port,
+            baudrate=self.baudrate,
+            # method="rtu",
+            stopbits=1,
+            bytesize=8,
+        )
 
     async def connect(self) -> None:
-        """Connect.
+        """Connect. Will create a new client if required."""
+        if not self.client:
+            self.client = self._new_client()
 
-        https://pymodbus.readthedocs.io/en/latest/source/example/async_asyncio_serial_client.html
-
-        """
-        if self.client and self.client.async_connected:
-            return
-        url = urlparse(f"{self.port}")
-
-        if not url.netloc:
-            # Cannot run from a coroutine currently
-            # https://github.com/riptideio/pymodbus/pull/658/files#r718775308
-            _LOGGER.debug("Port: %s pymodbus %s", self.port, version.short())
-
-            self.client = AsyncModbusSerialClient(
-                port=self.port,
-                baudrate=self.baudrate,
-                # method="rtu",
-                stopbits=1,
-                bytesize=8,
-            )
-            # _loop, client = msc  # pylint: disable=unpacking-non-sequence
-
-            # Alternative interim...
-            # client = AsyncioModbusSerialClient(
-            #     port=self.port,
-            #     protocol_class=ModbusClientProtocol,
-            #     framer=AsyncModbusSerialClient._framer(method="rtu"),
-            #     loop=asyncio.get_running_loop(),
-            #     baudrate=self.baudrate,
-            #     stopbits=STOPBITS_ONE,
-            #     bytesize=8,
-            # )
-        else:
-            _LOGGER.debug(
-                "Host: %s : %s pymodbus %s", url.hostname, url.port, version.short()
-            )
-            self.client = AsyncModbusTcpClient(
-                host=url.hostname,
-                port=url.port or 502,
-                # protocol_class=ModbusClientProtocol,
-                # loop=asyncio.get_running_loop(),
-            )
-
-        await self.client.connect()
+        if not self.client.async_connected:
+            await self.client.connect()
 
         if not self.client.async_connected:
             raise ConnectionError
 
-        # try:
-        #     client.protocol._timeout = self.timeout
-        # except AttributeError as err:
-        #     _LOGGER.warning("%s", err)
-
-        # self.client = client. .protocol
-
     async def write_register(self, *, address: int, value: int) -> bool:
         """Write to a register - Sunsynk supports modbus function 0x10."""
+        await self.connect()
         try:
-            w_r = await self.client.write_registers(
-                address=address, values=(value,), slave=self.server_id
+            res = await self.client.write_registers(  # type:ignore
+                address=address, values=[value], slave=self.server_id
             )
-            if w_r.function_code < 0x80:  # test that we are not an error
+            if res.function_code < 0x80:  # test that we are not an error
                 return True
             _LOGGER.error("failed to write register %s=%s", address, value)
         except asyncio.TimeoutError:
@@ -97,11 +68,12 @@ class pySunsynk(Sunsynk):  # pylint: disable=invalid-name
 
     async def read_holding_registers(self, start: int, length: int) -> Sequence[int]:
         """Read a holding register."""
-        res = await self.client.read_holding_registers(
+        await self.connect()
+        res = await self.client.read_holding_registers(  # type:ignore
             address=start, count=length, slave=self.server_id
         )
         if res.function_code >= 0x80:  # test that we are not an error
-            raise Exception(  # pylint: disable=broad-exception-raised
+            raise IOError(
                 f"failed to read register {start} - function code: {res.function_code}"
             )
         return res.registers