+++ tests/hass-addon-sunsynk-multi/__init__.py
@@ -1,2 +1,0 @@
-"""Add-on tests."""
-MOD_FOLDER = "hass-addon-sunsynk-multi"