+++ tests/hass_addon_sunsynk_multi/test_run.py
@@ -7,21 +7,15 @@ from types import ModuleType
 import pytest
 
 from tests.conftest import import_module
+from tests.hass_addon_sunsynk_multi import MOD_FOLDER
 
 _LOGGER = logging.getLogger(__name__)
 
 
 @pytest.fixture
-def mod_folder() -> str:
-    """Return the folder from __init__."""
-    init = import_module("__init__", Path(__file__).parent)
-    return init.MOD_FOLDER
-
-
-@pytest.fixture
-def run(mod_folder) -> ModuleType:
+def run() -> ModuleType:
     """Import the module."""
-    return import_module("run", mod_folder)
+    return import_module("run", MOD_FOLDER)
 
 
 def test_run(run):
@@ -29,18 +23,8 @@ def test_run(run):
     assert not run.STATES
     assert not run.OPT.mqtt_host
 
-    # testargs = ["run.py", "host1", "passw"]
-    # with patch.object(run.sys, "argv", testargs):
-    #     run.startup()
-    # assert run.STATES
-    # assert run.OPT.mqtt_host == "host1"
-    # assert run.OPT.mqtt_password == "passw"
 
-    # run.STATES.clear()
-    # run.OPT.mqtt_host = ""
-
-
-def test_versions(run, mod_folder):
+def test_versions(run):
     """Test versions.
 
     config.json - contains the HASS addon version
@@ -52,6 +36,7 @@ def test_versions(run, mod_folder):
         txt = Path(filename).read_text()
         res = re.compile(regex).search(txt)
         assert res, f"version not found in {filename}"
+        _LOGGER.warning("\t%s\t%s", res.group(1), filename)
         return res.group(1)
 
     v_setup = _get_version(
@@ -60,12 +45,12 @@ def test_versions(run, mod_folder):
     )
 
     v_docker = _get_version(
-        filename=f"{mod_folder}/Dockerfile",
+        filename=f"{MOD_FOLDER}/Dockerfile",
         regex=r"sunsynk(?:\[[^\]]+\])?==([0-9.]+)",
     )
 
     v_config = _get_version(
-        filename=f"{mod_folder}/config.yaml",
+        filename=f"{MOD_FOLDER}/config.yaml",
         regex=r"version: .+-(.+)",
     )